requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
cssselect==1.1.0
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
flaskwebgui==0.3.5
frozenlist==1.3.0
idna==3.3
importlib-metadata==4.11.3
lxml==4.8.0
multidict==6.0.2
Pillow==9.1.0
pycodestyle==2.8.0
pyee==8.2.2
PyJWT==2.3.0
pyppeteer==1.0.2
pyquery==1.4.3
pytils==0.4.1
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
websockets==10.2
whichcraft==0.6.1
yarl==1.7.2
zipp==3.8.0
